+++ src/packerlicious/provisioner.py
@@ -92,6 +92,46 @@ class File(PackerProvisioner):
     }
 
 
+class PowerShell(PackerProvisioner):
+    """
+    PowerShell Provisioner
+    https://www.packer.io/docs/provisioners/powershell.html
+    """
+    resource_type = "powershell"
+
+    # PowerShell Template Variables
+    Path = TemplateVar("Path")
+    Vars = TemplateVar("Vars")
+
+    # PowerShell Environment Variables
+    PackerBuildName = EnvVar("PACKER_BUILD_NAME")
+    PackerBuilderType = EnvVar("PACKER_BUILDER_TYPE")
+    PackerHttpAddr = EnvVar("PACKER_HTTP_ADDR")
+
+    props = {
+        'inline': ([basestring], False),
+        'script': (basestring, False),
+        'scripts': ([basestring], False),
+        'binary': (validator.boolean, False),
+        'environment_vars': ([basestring], False),
+        'execute_command': (basestring, False),
+        'elevated_user': (basestring, False),
+        'elevated_password': (basestring, False),
+        'remote_path': (basestring, False),
+        'skip_clean': (validator.boolean, False),
+        'start_retry_timeout': (basestring, False),
+        'valid_exit_codes': ([int], False),
+    }
+
+    def validate(self):
+        conds = [
+            'inline',
+            'script',
+            'scripts'
+        ]
+        validator.exactly_one(self.__class__.__name__, self.properties, conds)
+
+
 class PuppetMasterless(PackerProvisioner):
     """
     Puppet Masterless
@@ -134,7 +174,6 @@ class PuppetServer(PackerProvisioner):
         'puppet_server': (basestring, False),
         'staging_dir': (basestring, False),
     }
-
 
 
 class SaltMasterless(PackerProvisioner):